pyproject.toml
[tool.poetry]
name = "algo-mvp"
version = "0.3.0"
description = "Lean backtest + paper-trade stack for Mike (Dock108)"
authors = ["Dock108 <mike@dock108.ai>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10"
backtrader = "^1.9.78.123"
vectorbt = "^0.26.0"
streamlit = "^1.45.0"
alpaca-trade-api = "^3.2.0"
sqlalchemy = "^2.0.40"
pydantic = "^2.8.2"
tenacity = "^8.5.0"
pyarrow = "^17.0.0"
pendulum = "^3.0.0"
pyyaml = "^6.0.1"
requests = "^2.32.3"
plotly = "^5.24.0"
rich = "^13.7.1"
tqdm = "^4.66.4"

[tool.poetry.group.dev.dependencies]
responses = "^0.25.7"
pytest-cov = "^6.1.1"
pytest-mock = "^3.14.0"
pre-commit = "^4.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"
addopts = "--cov=src --cov-report=term-missing"

[tool.coverage.run]
source = ["src"]
omit = [
    "*/__main__.py",
    # "src/algo_mvp/data/alpaca.py", # Removed - too broad
    # "src/algo_mvp/data/tradovate.py", # Removed - too broad
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG is True:",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    'if __name__ == "__main__":', # Use single quotes to simplify
    "if TYPE_CHECKING:",
    "@(abc\\.)?abstractmethod", # Escape the backslash for the dot
]
ignore_errors = true
fail_under = 80 # Keep overall fail under here

# Optional: Define specific path coverage if needed later
# [tool.coverage.paths]
# source = ["src", ".tox/*/lib/python*/site-packages"]

# Optional: Fine-grained report settings
# [tool.coverage.html]
# directory = "coverage_html_report"
